refactor: remove extractCustomerId method
