+++ src/main/java/com/swifty/bank/server/api/service/impl/AuthenticationApiServiceImpl.java
@@ -2,17 +2,18 @@ package com.swifty.bank.server.api.service.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.swifty.bank.server.api.controller.dto.TokenDto;
 import com.swifty.bank.server.api.controller.dto.auth.request.JoinRequest;
 import com.swifty.bank.server.api.service.AuthenticationApiService;
 import com.swifty.bank.server.api.service.dto.ResponseResult;
 import com.swifty.bank.server.api.service.dto.Result;
 import com.swifty.bank.server.core.common.authentication.Auth;
-import com.swifty.bank.server.core.common.authentication.service.impl.AuthenticationServiceImpl;
+import com.swifty.bank.server.core.common.authentication.dto.TokenDto;
+import com.swifty.bank.server.core.common.authentication.service.AuthenticationService;
+import com.swifty.bank.server.core.common.utils.JwtUtil;
 import com.swifty.bank.server.core.common.utils.RedisUtil;
 import com.swifty.bank.server.core.common.utils.StringUtil;
 import com.swifty.bank.server.core.domain.customer.Customer;
-import com.swifty.bank.server.core.domain.customer.service.impl.CustomerServiceImpl;
+import com.swifty.bank.server.core.domain.customer.service.CustomerService;
 import com.swifty.bank.server.exception.AuthenticationException;
 import com.swifty.bank.server.exception.StoredAuthValueNotExistException;
 import java.util.HashMap;
@@ -28,8 +29,8 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 public class AuthenticationApiServiceImpl implements AuthenticationApiService {
-    private final CustomerServiceImpl customerService;
-    private final AuthenticationServiceImpl authenticationService;
+    private final CustomerService customerService;
+    private final AuthenticationService authenticationService;
     private final RedisUtil redisUtil;
 
     @Override
@@ -102,7 +103,7 @@ public class AuthenticationApiServiceImpl implements AuthenticationApiService {
             Map<String, String> map = mapper.readValue(body, Map.class);
             refreshToken = map.get("RefreshToken");
 
-            uuid = authenticationService.extractCustomerId(refreshToken);
+            uuid = UUID.fromString(JwtUtil.getClaimByKey(refreshToken, "customerId").toString());
         } catch (JsonProcessingException e) {
             return new ResponseResult<>(
                     Result.FAIL,
@@ -151,7 +152,7 @@ public class AuthenticationApiServiceImpl implements AuthenticationApiService {
     public ResponseResult<?> logout(String token) {
         UUID uuid;
         try {
-            uuid = authenticationService.extractCustomerId(token);
+            uuid = UUID.fromString(JwtUtil.getClaimByKey(token, "customerId").toString());
         } catch (AuthenticationException e) {
             return new ResponseResult<>(
                     Result.FAIL,
@@ -162,7 +163,7 @@ public class AuthenticationApiServiceImpl implements AuthenticationApiService {
 
         try {
             authenticationService.logout(uuid);
-            return new ResponseResult<>(Result.SUCCESS, "[INFO] user " + uuid.toString() + " logged out", null);
+            return new ResponseResult<>(Result.SUCCESS, "[INFO] user " + uuid + " logged out", null);
         } catch (StoredAuthValueNotExistException e) {
             return new ResponseResult<>(
                     Result.FAIL,
@@ -176,7 +177,7 @@ public class AuthenticationApiServiceImpl implements AuthenticationApiService {
     public ResponseResult<?> signOut(String token) {
         UUID uuid;
         try {
-            uuid = authenticationService.extractCustomerId(token);
+            uuid = UUID.fromString(JwtUtil.getClaimByKey(token, "customerId").toString());
         } catch (AuthenticationException e) {
             return new ResponseResult<>(
                     Result.FAIL,
@@ -207,7 +208,7 @@ public class AuthenticationApiServiceImpl implements AuthenticationApiService {
         Map<String, Object> result = new HashMap<>();
 
         try {
-            TokenDto tokens = authenticationService.generateTokenDtoWithCustomer(customer);
+            TokenDto tokens = authenticationService.generateTokenDto(customer);
             authenticationService.saveRefreshTokenInDataSources(tokens.getRefreshToken());
             result.put("token", tokens);
             return new ResponseResult<>(
@@ -225,7 +226,7 @@ public class AuthenticationApiServiceImpl implements AuthenticationApiService {
     }
 
     private boolean isValidatedRefreshToken(String token) {
-        UUID uuid = authenticationService.extractCustomerId(token);
+        UUID uuid = UUID.fromString(JwtUtil.getClaimByKey(token, "customerId").toString());
 
         Auth previousAuth = redisUtil.getRedisAuthValue(uuid.toString());
         if (previousAuth == null) {
