+++ src/main/java/com/swifty/bank/server/core/common/authentication/service/AuthenticationService.java
@@ -1,13 +1,13 @@
 package com.swifty.bank.server.core.common.authentication.service;
 
-import com.swifty.bank.server.api.controller.dto.TokenDto;
 import com.swifty.bank.server.core.common.authentication.Auth;
+import com.swifty.bank.server.core.common.authentication.dto.TokenDto;
 import com.swifty.bank.server.core.domain.customer.Customer;
 import java.util.Optional;
 import java.util.UUID;
 
 public interface AuthenticationService {
-    TokenDto generateTokenDtoWithCustomer(Customer customer);
+    TokenDto generateTokenDto(Customer customer);
 
     void logout(UUID uuid);
