+++ src/main/java/com/swifty/bank/server/core/common/authentication/service/impl/AuthenticationServiceImpl.java
@@ -1,7 +1,7 @@
 package com.swifty.bank.server.core.common.authentication.service.impl;
 
-import com.swifty.bank.server.api.controller.dto.TokenDto;
 import com.swifty.bank.server.core.common.authentication.Auth;
+import com.swifty.bank.server.core.common.authentication.dto.TokenDto;
 import com.swifty.bank.server.core.common.authentication.repository.AuthRepository;
 import com.swifty.bank.server.core.common.authentication.service.AuthenticationService;
 import com.swifty.bank.server.core.common.utils.DateUtil;
@@ -33,7 +33,27 @@ public class AuthenticationServiceImpl implements AuthenticationService {
     private int refreshTokenExpiration;
 
     @Override
-    public TokenDto generateTokenDtoWithCustomer(Customer customer) {
+    public String createAccessToken(Customer customer) {
+        Claims claims = Jwts.claims();
+        Date expiration = DateUtil.millisToDate(DateUtil.now().getTime() + accessTokenExpiration * 1000L);
+
+        claims.setSubject("AccessToken");
+        claims.put("customerId", customer.getId());
+        return JwtUtil.generateToken(claims, expiration);
+    }
+
+    @Override
+    public String createRefreshToken(Customer customer) {
+        Claims claims = Jwts.claims();
+        Date expiration = DateUtil.millisToDate(DateUtil.now().getTime() + refreshTokenExpiration * 1000L);
+
+        claims.setSubject("RefreshToken");
+        claims.put("customerId", customer.getId());
+        return JwtUtil.generateToken(claims, expiration);
+    }
+
+    @Override
+    public TokenDto generateTokenDto(Customer customer) {
         return new TokenDto(createAccessToken(customer), createRefreshToken(customer));
     }
 
@@ -68,35 +88,11 @@ public class AuthenticationServiceImpl implements AuthenticationService {
     public Optional<Auth> findAuthByUuid(UUID uuid) {
         return authRepository.findAuthByUuid(uuid);
     }
-
-    @Override
-    public String createAccessToken(Customer customer) {
-        Claims claims = Jwts.claims();
-        Date expiration = DateUtil.millisToDate(DateUtil.now().getTime() + accessTokenExpiration * 1000L);
-
-        claims.setSubject("AccessToken");
-        claims.put("customerId", customer.getId());
-        return JwtUtil.generateToken(claims, expiration);
-    }
-
-    @Override
-    public String createRefreshToken(Customer customer) {
-        Claims claims = Jwts.claims();
-        Date expiration = DateUtil.millisToDate(DateUtil.now().getTime() + refreshTokenExpiration * 1000L);
-
-        claims.setSubject("RefreshToken");
-        claims.put("customerId", customer.getId());
-        return JwtUtil.generateToken(claims, expiration);
-    }
-
-    public UUID extractCustomerId(String jwt) {
-        return UUID.fromString(JwtUtil.getClaimByKey(jwt, "customerId").toString());
-    }
-
+    
     @Override
     @Transactional
-    public void saveRefreshTokenInDataSources(String token) {
-        UUID uuid = UUID.fromString(JwtUtil.getClaimByKey(token, "customerId").toString());
+    public void saveRefreshTokenInDataSources(String jwt) {
+        UUID uuid = UUID.fromString(JwtUtil.getClaimByKey(jwt, "customerId").toString());
 
         Auth previousAuth = redisUtil.getRedisAuthValue(uuid.toString());
         if (previousAuth == null) {
@@ -106,10 +102,10 @@ public class AuthenticationServiceImpl implements AuthenticationService {
         Auth newAuth;
 
         if (previousAuth != null) {
-            previousAuth.updateAuthContent(token);
+            previousAuth.updateAuthContent(jwt);
             newAuth = previousAuth;
         } else {
-            newAuth = new Auth(uuid, token);
+            newAuth = new Auth(uuid, jwt);
             authRepository.save(newAuth);
         }
         redisUtil.saveAuthRedis(uuid.toString(), newAuth);
